authentication completed

+++ src/DDD/Identity.Infrastructure/Services/Concrete/AuthService.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Generic;
+﻿using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Reflection.Metadata.Ecma335;
@@ -12,7 +13,7 @@ using Microsoft.Extensions.Options;
 
 namespace Identity.Infrastructure.Services.Concrete
 {
-    public class AuthService:IAuthService
+    public class AuthService : IAuthService
     {
         private readonly List<Client> _clients;
         private readonly ITokenService _tokenService;
@@ -31,17 +32,17 @@ namespace Identity.Infrastructure.Services.Concrete
             _userRoleService = userRoleService;
             _userRefreshTokenService = userRefreshTokenService;
         }
-        
+
         public async Task<GenericResult<UserTokenDto>> CreateUserTokenAsync(LoginRequest request)
         {
             var user = await _userService.FindUserByEmailAsync(request.Email);
-            if (user==null)
+            if (user == null)
             {
                 ErrorResult error = new("Girilen adrese ait bir hesap bulunamadı.");
                 return GenericResult<UserTokenDto>.ErrorResponse(error, statusCode: 400);
             }
 
-            if (!_userService.CheckPassword(user,request.Password))
+            if (!_userService.CheckPassword(user, request.Password))
             {
                 ErrorResult error = new("Email veya şifre hatalıdır.");
                 return GenericResult<UserTokenDto>.ErrorResponse(error, statusCode: 400);
@@ -51,10 +52,10 @@ namespace Identity.Infrastructure.Services.Concrete
             var token = _tokenService.CreateUserToken(user, roles);
 
             var userRefreshToken = await _userRefreshTokenService.FindAsync(x => x.RefreshToken == token.RefreshToken);
-            if (userRefreshToken==null)
+            if (userRefreshToken == null)
             {
                 await _userRefreshTokenService.InsertAsync(new UserRefreshToken
-                    { UserId = user.Id, RefreshToken = token.RefreshToken });
+                { UserId = user.Id, RefreshToken = token.RefreshToken,Expiration = token.RefreshTokenExpiration});
             }
             else
             {
@@ -62,44 +63,51 @@ namespace Identity.Infrastructure.Services.Concrete
                 userRefreshToken.Expiration = token.RefreshTokenExpiration;
                 await _userRefreshTokenService.UpdateAsync(userRefreshToken);
             }
-            
-            return GenericResult<UserTokenDto>.SuccessResponse(token,(int)HttpStatusCode.OK);
+
+            return GenericResult<UserTokenDto>.SuccessResponse(token, (int)HttpStatusCode.OK);
         }
 
         public async Task<GenericResult<UserTokenDto>> CreateTokenByRefreshToken(string refreshToken)
         {
-            var userRefreshToken = await _userRefreshTokenService.FindAsync(x => x.RefreshToken == refreshToken);
+            var userRefreshToken = await _userRefreshTokenService.FindAsync(x => x.RefreshToken == refreshToken && x.Expiration > DateTime.Now);
+
+            if (userRefreshToken == null)
+                return GenericResult<UserTokenDto>.ErrorResponse(new ErrorResult("RefreshToken süresi doldu."), (int)HttpStatusCode.BadRequest);
 
             var user = await _userService.FindAsync(userRefreshToken.UserId);
             var roles = await _userRoleService.GetRolesByUserId(user.Id);
             var token = _tokenService.CreateUserToken(user, roles);
-            
+
             //update
             userRefreshToken.RefreshToken = token.RefreshToken;
             userRefreshToken.Expiration = token.RefreshTokenExpiration;
             await _userRefreshTokenService.UpdateAsync(userRefreshToken);
-            
-            return GenericResult<UserTokenDto>.SuccessResponse(token,(int)HttpStatusCode.OK);
+
+            return GenericResult<UserTokenDto>.SuccessResponse(token, (int)HttpStatusCode.OK);
         }
 
         public async Task<GenericResult<string>> RevokeRefreshToken(string refreshToken)
         {
             var userRefreshToken = await _userRefreshTokenService.FindAsync(x => x.RefreshToken == refreshToken);
+
+            if (userRefreshToken == null)
+                return GenericResult<string>.ErrorResponse(new ErrorResult("RefreshToken bulunamadı."), (int)HttpStatusCode.BadRequest);
+
             await _userRefreshTokenService.DeleteAsync(userRefreshToken);
-            
-            return GenericResult<string>.SuccessResponse("User refresh token deleted",200);
+
+            return GenericResult<string>.SuccessResponse("User refresh token deleted", 200);
         }
 
         public GenericResult<ClientTokenDto> CreateClientToken(ClientTokenRequest request)
         {
             var client = _clients.SingleOrDefault(x => x.Id == request.ClientId && x.Secret == request.ClientSecret);
 
-            if (client==null)
-                return GenericResult<ClientTokenDto>.ErrorResponse(new ErrorResult("Client bilgileri hatalı!"),(int)HttpStatusCode.BadRequest);
+            if (client == null)
+                return GenericResult<ClientTokenDto>.ErrorResponse(new ErrorResult("Client bilgileri hatalı!"), (int)HttpStatusCode.BadRequest);
 
             var token = _tokenService.CreateClientToken(client);
-            
-            return GenericResult<ClientTokenDto>.SuccessResponse(token,(int)HttpStatusCode.OK);
+
+            return GenericResult<ClientTokenDto>.SuccessResponse(token, (int)HttpStatusCode.OK);
         }
     }
 }
